longevity -> durability
The system is actually impossible to migrate away from and still drives the FireTV software remote, so it's hard to say it doesn't have longevity. But it's not the kind of longevity you prefer to have. A mentor of mine once told me, the goal of software is to be rewritten and reabsorbed.